Fix `ginkgo` import chain in `manager` cmd
