+++ internal/webhook/v1alpha1/bmcsettings_webhook_test.go
@@ -6,7 +6,6 @@ package v1alpha1
 import (
 	"fmt"
 
-	"github.com/ironcore-dev/metal-operator/internal/controller"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	v1 "k8s.io/api/core/v1"
@@ -45,9 +44,6 @@ var _ = Describe("BMCSettings Webhook", func() {
 	AfterEach(func() {
 		By("Deleting BMCSettings")
 		Expect(k8sClient.DeleteAllOf(ctx, &metalv1alpha1.BMCSettings{})).To(Succeed())
-
-		By("Ensuring clean state")
-		controller.EnsureCleanState()
 	})
 
 	Context("When creating or updating BMCSettings under Validating Webhook", func() {
